+++ ali/fabric/src/main/java/com/yanny/ali/fabric/plugin/FabricPlugin.java
@@ -1,0 +1,21 @@
+package com.yanny.ali.fabric.plugin;
+
+import com.yanny.ali.api.AliEntrypoint;
+import com.yanny.ali.api.IPlugin;
+import com.yanny.ali.api.IServerRegistry;
+import net.fabricmc.fabric.impl.recipe.ingredient.CustomIngredientImpl;
+import org.jetbrains.annotations.NotNull;
+
+@AliEntrypoint
+public class FabricPlugin implements IPlugin {
+    @NotNull
+    @Override
+    public String getModId() {
+        return "fabric";
+    }
+
+    @Override
+    public void registerServer(IServerRegistry registry) {
+        registry.registerIngredientTooltip(CustomIngredientImpl.class, FabricIngredientTooltipUtils::getCustomIngredientTooltip);
+    }
+}