+++ ali/neoforge/src/main/java/com/yanny/ali/neoforge/plugin/NeoForgePlugin.java
@@ -21,12 +21,20 @@ import com.yanny.ali.plugin.glm.ILootTableIdConditionPredicate;
 import net.minecraft.resources.RegistryOps;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.neoforged.neoforge.common.BasicItemListing;
 import net.neoforged.neoforge.common.ItemAbility;
+import net.neoforged.neoforge.common.crafting.BlockTagIngredient;
+import net.neoforged.neoforge.common.crafting.CompoundIngredient;
+import net.neoforged.neoforge.common.crafting.DataComponentIngredient;
+import net.neoforged.neoforge.common.crafting.DifferenceIngredient;
+import net.neoforged.neoforge.common.crafting.ICustomIngredient;
+import net.neoforged.neoforge.common.crafting.IntersectionIngredient;
 import net.neoforged.neoforge.common.loot.*;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.Unmodifiable;
 import org.slf4j.Logger;
 import oshi.util.tuples.Pair;
@@ -49,12 +57,25 @@ public class NeoForgePlugin implements IPlugin {
         registry.registerConditionTooltip(CanItemPerformAbility.class, NeoForgePlugin::getCanToolPerformActionTooltip);
         registry.registerConditionTooltip(LootTableIdCondition.class, NeoForgePlugin::getLootTableIdTooltip);
 
+        registry.registerIngredientUnwrapper(NeoForgePlugin::unwrapCustomIngredient);
+
+        registry.registerValueTooltip(CompoundIngredient.class, NeoForgeIngredientTooltipUtils::getCompoundIngredientTooltip);
+        registry.registerValueTooltip(DifferenceIngredient.class, NeoForgeIngredientTooltipUtils::getDifferenceIngredientTooltip);
+        registry.registerValueTooltip(IntersectionIngredient.class, NeoForgeIngredientTooltipUtils::getIntersectionIngredientTooltip);
+        registry.registerValueTooltip(DataComponentIngredient.class, NeoForgeIngredientTooltipUtils::getDataComponentIngredientTooltip);
+        registry.registerValueTooltip(BlockTagIngredient.class, NeoForgeIngredientTooltipUtils::getBlockTagIngredientTooltip);
+
         registry.registerItemListing(BasicItemListing.class, NeoForgePlugin::getBasicItemListingNode);
         registry.registerItemListingCollector(BasicItemListing.class, NeoForgePlugin::collectBasicItemListingItems);
 
         registry.registerLootModifiers(NeoForgePlugin::registerLootModifiers);
 
         registry.registerValueTooltip(ItemAbility.class, NeoForgePlugin::getItemAbilityTooltip);
+    }
+
+    @Nullable
+    private static ICustomIngredient unwrapCustomIngredient(Ingredient ingredient) {
+        return ingredient.isCustom() ? ingredient.getCustomIngredient() : null;
     }
 
     @NotNull